Adds support for round IDs + related README changes
Adds the current round ID as a footer in admin/new round notifications. These changes require slight
modifications to the redbot subsystem to function, please see the updated code contained in the README for more
information. This is a non-breaking change so if the subsystem is not modified the footer simply will not appear.